requirements.txt
Django==1.9.1

djangorestframework==3.3.3
django-oauth-toolkit==0.10.0
django-rest-swagger==0.3.4
oauthlib==1.0.3
psycopg2==2.6.1
